Decompose public_reservations into focused helper modules
